Reorganize the package.
Move the basic functions for every flavour of optic
into a dedicated module.

The split between internal modules and "external" modules
is maintained so that the `Optics.Internal.Subtyping`
module can depend on all internal modules, because it needs
access to all the tag types. In turn, each of the external
modules depends on the `Optics.Internal.Subtyping` module.