Change abstract interface
From

    newtype Optic k i o s t a b =
      Optic { forall p. Constraints k p => p i s t -> p o s t }

to

    newtype Optic k is s t a b =
      Optic { forall p j. Constraints k p => p j s t -> p (Curry is j) s t }

Worth noticing, that unindexed eliminators accept optics with any
indices. This is choice between
- be more liberate what we can accept; and
- guide to not use indexed optics when you don't need them.

Therefore we have formation:

    type Lens s t a b = Optic A_Lens '[] s t a b

and constructors use the type alias (there cannot be any indices)

    sets :: ((a -> b) -> s -> t) -> Setter s t a b

but eliminator uses `Is k ... => Optic k ...`:

    over :: Is k A_Setter => Optic k is s t a b
         -> (a -> b) -> s -> t

Also, conversion function doesn't use alias, because it forces indices
to be '[]:

    toSetter :: Is k A_Setter
             -> Optic k is s t a b
             -> Optic A_Setter is s t a b

In general, refactoring is quite direct and completely type-error
driven.