Add two tests about unindexing
At least in simple cases

- unIx doesn't introduce visible overhead (`constN` expands)
- using indexed optic with e.g. over directly
  will introduce the overhead of only calling `imap`
  (and not `fmap`).

The latest point is why `lens` has Conjoined.
As the optic is a part of `FunctorWithIndex`, it has definition

    imapped :: IndexedSetter i (f a) (f b) a b
    imapped = conjoined mapped (isets imap)

i.e. if you don't use indices, use `fmap` not `imap`.

Final note: inspection-testing doesn't believe
that with GHC-8.0.2 they are the same; but the Core looks fine.